+++ pyrit/score/__init__.py
@@ -5,7 +5,7 @@ from pyrit.models import Score, ScoreType  # noqa: F401
 
 from pyrit.score.scorer import Scorer
 
-from pyrit.score.hitl_scorer import HumanInTheLoopScorer
+from pyrit.score.human_in_the_loop_scorer import HumanInTheLoopScorer
 from pyrit.score.self_ask_category_scorer import SelfAskCategoryScorer, ContentClassifierPaths
 from pyrit.score.self_ask_likert_scorer import SelfAskLikertScorer, LikertScalePaths
 from pyrit.score.self_ask_true_false_scorer import SelfAskTrueFalseScorer, TrueFalseQuestionPaths
