+++ tests/score/test_hitl.py
@@ -15,7 +15,7 @@ from pyrit.models.prompt_request_piece import PromptRequestPiece
 
 
 from pyrit.models.score import Score
-from pyrit.score.hitl_scorer import HumanInTheLoopScorer
+from pyrit.score.human_in_the_loop_scorer import HumanInTheLoopScorer
 from pyrit.score.substring_scorer import SubStringScorer
 from tests.mocks import get_image_request_piece, get_memory_interface
